Pontas soltas usuario e alunos

+++ controllers/usuario/usuario.php
@@ -11,11 +11,6 @@ class Usuario extends \Libs\Controller {
 		'send'		=> 'Usuário'
 	];
 
-
-
-
-
-
 	function __construct() {
 		parent::__construct();
 		\Util\Auth::handLeLoggin();
@@ -34,15 +29,28 @@ class Usuario extends \Libs\Controller {
 	}
 
 	public function listagem($dados_linha){
+		if(empty($dados_linha)){
+			return false;
+		}
+
+		foreach ($this->model->load_active_list('hierarquia') as $indice => $hierarquia) {
+			$hierarquias[$hierarquia['id']] = $hierarquia['nome'];
+		};
+
 		foreach ($dados_linha as $indice => $linha) {
 			if($linha['super_admin'] != 1){
+
 				$retorno_linhas[] = [
 					"<td class='sorting_1'>{$linha['id']}</td>",
 	        		"<td>{$linha['email']}</td>",
-	        		"<td>{$linha['hierarquia']}</td>",
+	        		"<td>{$hierarquias[$linha['hierarquia']]}</td>",
 	        		"<td>" . $this->view->default_buttons_listagem($linha['id']) . "</td>"
 				];
 			}
+		}
+
+		if(!isset($retorno_linhas)){
+			return false;
 		}
 
 		$this->view->linhas_datatable = $retorno_linhas;
@@ -50,6 +58,13 @@ class Usuario extends \Libs\Controller {
 
 	public function editar($id) {
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "editar");
+
+		if(empty($this->model->db->select("SELECT id FROM {$this->modulo['modulo']} WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
+
 		$this->view->cadastro = $this->model->full_load_by_id('usuario', $id[0])[0];
 		$this->view->hierarquia_list = $this->model->load_active_list('hierarquia');
 		$this->view->render($this->modulo['modulo'] . '/editar/editar');
@@ -57,6 +72,13 @@ class Usuario extends \Libs\Controller {
 
 	public function visualizar($id){
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "visualizar");
+
+
+		if(empty($this->model->db->select("SELECT id FROM {$this->modulo['modulo']} WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
 
 		$this->view->cadastro = $this->model->full_load_by_id('usuario', $id[0])[0];
 		$this->view->hierarquia_list = $this->model->load_active_list('hierarquia');
@@ -88,6 +110,13 @@ class Usuario extends \Libs\Controller {
 	public function update($id) {
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "editar");
 
+
+		if(empty($this->model->db->select("SELECT id FROM {$this->modulo['modulo']} WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
+
 		$update_db = carregar_variavel('usuario');
 
 		if(empty($update_db['senha'])){
@@ -107,6 +136,18 @@ class Usuario extends \Libs\Controller {
 
 	public function delete($id) {
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "deletar");
+
+		if(empty($this->model->db->select("SELECT id FROM {$this->modulo['modulo']} WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
+
+		if($_SESSION['usuario']['id'] == $id[0]){
+			$this->view->alert_js('Você não pode excluir seu proprio usuário!!!', 'erro');
+			header('location: ' . URL . 'usuario');
+			exit;
+		}
 
 		$retorno = $this->model->delete('usuario', $id[0]);
